+++ src/app/application-modules/dashboard/components/dashboard-details/dashboard-details.component.ts
@@ -4,9 +4,9 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { TranslateService } from '@ngx-translate/core';
 import { combineLatest, Subscription } from 'rxjs';
 import { filter } from 'rxjs/operators';
-import { Forecast } from '../../../../model/forecast';
-import { ForecastDay } from '../../../../model/forecast-day';
-import { Hour } from '../../../../model/hour';
+import { Forecast } from '../../../../model/weather/forecast';
+import { ForecastDay } from '../../../../model/weather/forecast-day';
+import { Hour } from '../../../../model/weather/hour';
 import { MenuService } from '../../../../shared/shared.module';
 import { Utils } from '../../../../shared/utils';
 import { slideInOutAnimation } from './slide-in-out';
