+++ src/app/application-modules/user/components/signin/signin.component.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/styles';
+@import '../../../../styles/styles';
 
 .signin-content {
   display: flex;
