+++ src/app/application-modules/user/components/signin/signin.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 import { TranslateService } from '@ngx-translate/core';
-import { AuthService } from '../../../shared/service/auth.service';
+import { AuthService } from '../../../../shared/shared.module';
 
 @Component({
     selector: 'app-signin',
