+++ src/app/application-modules/user/components/signup/signup.component.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/styles';
+@import '../../../../styles/styles';
 
 .signup-content {
   display: flex;
