+++ src/app/model/weather/condition.ts
