+++ src/app/model/weather/day.ts
