+++ src/app/model/weather/forecast-day.ts
