+++ src/app/model/weather/forecast.ts
