+++ src/app/model/weather/hour.ts
