+++ src/app/shared/component/search-location/search-location.component.ts
@@ -8,7 +8,7 @@ import {
 import { MatAutocompleteTrigger } from '@angular/material/autocomplete';
 import { faLocationArrow, faTimes } from '@fortawesome/free-solid-svg-icons';
 import { Observable, Observer } from 'rxjs';
-import { Location } from '../../../model/location';
+import { Location } from '../../../model/weather/location';
 import { WeatherService } from '../../service/weather.service';
 
 @Component({
